pyproject.toml
[project]
name = "easyblog"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Erfan Kheyrollahi", email = "ekm507@gmail.com" },
    { name = "Amir Aref", email = "amiraref808@gmail.com" }
]
dependencies = [
    "jdatetime>=5.0.0",
    "markdown-it-py>=3.0.0",
    "jinja2>=3.1.4",
]
readme = "README.md"
requires-python = ">= 3.11"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
